pyproject.toml
[project]
name = "budget-parser"
version = "0.1.0"
description = "CLI tool for classifying transactions into budget categories with a local LLM"
readme = "README.md"
authors = [
    { name = "Codex", email = "codex@openai.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.0.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "tqdm>=4.67.1",
    "typer>=0.17.4",
]

[project.scripts]
budget-parser = "budget_parser:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
